// examples/convex/tasks/convex/_generated/dataModel.d.ts
// Generated by @nimbus/codegen. Do not edit by hand.
import type { GenericId } from "convex/values";

export type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export type TableNames = "tasks";

export type Id<TableName extends string> = GenericId<TableName>;

type DocumentByTable = {
  "tasks": {
    _id: Id<"tasks">;
    _creationTime: number;
    _updateTime: number;
    "text": string;
    "completed": boolean;
    "createdAt": number;
  };
};

type IndexNamesByTable = {
  "tasks": "by_created_at";
};

export type Doc<TableName extends TableNames> = DocumentByTable[TableName];

export type DataModel = {
  [TableName in TableNames]: {
    document: DocumentByTable[TableName];
    fieldPaths: keyof DocumentByTable[TableName];
    indexes: IndexNamesByTable[TableName];
  };
};